perf(web): lazy load user quota editor
